added the li in array and extracting the {values git
git status
