+++ src/App.js
@@ -1,6 +1,5 @@
 import "./App.css";
 
-import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import Header from "./components/header/header.jsx";
